Create revvy_remote_pybleno.py
Create a pybleno based implementation of the revvy BLE controller.

After failed attempts with the bluezero library, I've looked for something else.

Why bluezero is a deadend:
bluezero is based on BlueZ 5.43, as that is the default version that comes with Raspbian. The problem however, is that BlueZ only introduced the support for LocalName in their DBus API (that bluezero is using to communicate to BlueZ), which allows for sending a ScanResponse for clients defining a human readable name for the BLE peripheral.
I've tried directly sending the OCF_LE_SET_SCAN_RESPONSE_DATA through the BT HCI, but it wasn't working out. Also, it is quite a hacky approach to use and bypass BlueZ the same time.
Bleno directly talks to the BT chip through a socket (BT sockets are native in python3), through HCI (Host Controller Interface). The python version is a simple port of the node.js version. As we already have positive experience with Bleno, I decided to give the port a try.